pyproject.toml
[tool.poetry]
name = "repl_python3_RoundTeemingBash"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
tweepy = "^4.4.0"
Pillow = "^9.0.0"
Flask = "^2.0.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
